requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
httpx==0.26.0
python-decouple==3.8
spacy==3.7.2
pymongo==4.6.1
pytest==7.3.1
pytest-asyncio==0.21.0